refactor(frontend): move env var editor to component
